Use check_node with multiple dispatch after changes in EMB.
Changes in EMB included using modeltype::EnergyModel instead of modeltyp::OperationalModel in check methods to catch calls to the check methods that are not specified in the other packages, e.g. for calls from InvestmentModels. This resulted in the method call being ambiguous when there existed a method check_node(n::Node, T, modeltype::InvestmentModel).